fix: dto에서 price type을 big demical에서 int로 변경

+++ src/main/kotlin/com/mss/assignment/dto/CheapestEachCategory.kt
@@ -1,17 +1,15 @@
 package com.mss.assignment.dto
 
-import java.math.BigDecimal
-
 data class CheapestEachCategory(
     val items: List<ProductDto>,
-    val totalPrice: BigDecimal,
+    val totalPrice: Int,
 ) {
     companion object {
         fun fromProductDtoList(products: List<ProductDto>): CheapestEachCategory {
             val totalPrice = products.sumOf { it.price }
             return CheapestEachCategory(
                 items = products,
-                totalPrice = totalPrice,
+                totalPrice = totalPrice.toInt(),
             )
         }
     }
